reorganization and first test
